Add async retries (#151)
Related to infinyon/fluvio#2361.

This PR adds a generic retry mechanism in the async context.
An async operation that needs to be retried must be represented in the `FnMut -> Future` form. Essentially, it means that need to provide a factory of futures.
Control over the number of retries and timings is done by `Iterator<Item=Duration>`.
Also, added `timeout` method that transforms a future into a time-limited future.