feat: enable socket keepalive by default (#211)
Added an ability to set socket `keepalive` property.

By default, all fluvio TCP connections will have set `keepalive=true`.  This is needed to prevent silently closing TCP connections by OS or network gateways when there is no traffic for an extended period of time. It often happens for consumers when there is no data on the topic for a period of more than one day. TCP session gets closed but the consumer does not determine that and keeps waiting for the data. For instance https://github.com/infinyon/roadmap/issues/85